requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-filter==25.1
asgiref==3.9.1
sqlparse==0.5.3
tzdata==2025.2
